Minor restructure to architecture. Most major logic will now be consolidated into the UserService class and main controller will now be the UserController. Might change the names of it soon

+++ src/main/java/com/lbycpd2/todoexp/restful/taskpackage/ParentTaskService.java
@@ -1,9 +1,7 @@
 package com.lbycpd2.todoexp.restful.taskpackage;
 
-import com.lbycpd2.todoexp.restful.userpackage.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,30 +9,24 @@ import java.util.Optional;
 @Service
 public class ParentTaskService {
     private final ParentTaskRepository repository;
-    private final UserRepository userRepository;
 
     @Autowired
-    public ParentTaskService(ParentTaskRepository repository, UserRepository userRepository) {
+    public ParentTaskService(ParentTaskRepository repository) {
         this.repository = repository;
-        this.userRepository = userRepository;
     }
 
-    // TODO: Get all parent tasks
     public List<ParentTask> getAllParentTasks(){
         return repository.findAll();
     }
-
-    // TODO: Get specific parent task
-    public Optional<ParentTask> getParentTask(Long parent_id){
-        return repository.findById(parent_id);
+    public ParentTask getParentTask(Long parent_id){
+        Optional<ParentTask> optionalPTask = repository.findById(parent_id);
+        return optionalPTask.orElseThrow(IllegalAccessError::new);
     }
 
-    // TODO: Delete parent task
     public void deleteParentTask(Long parent_id){
+        boolean exists = repository.findById(parent_id).isPresent();
+        if(!exists) throw new IllegalStateException("Parent task does not exist!");
         repository.deleteById(parent_id);
     }
-
-    // TODO: Update parent task
-
 
 }
